+++ CHANGELOG.md
@@ -4,11 +4,15 @@ Changements apportés au module `pf1-fr`.
 Le format est basé sur [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 Ce projet adhère aux principes du [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.11.0] - 2021-01-31
+## [1.11.0] - 2021-02-01
 ### Ajouté
 - Nouveau bestiaire (merci à gobbhile#41770)
 - Macro pour configurer les images (non-incluses)
 - Mise en évidence du jeton du combatant à qui c'est le tour (piste d'initiative)
+- Macro de MJ : afficher les effets actifs sur les PJs
+- Macro de MJ : lister les changements actifs sur un PJ
+- Macro de MJ : afficher les richesses des PJs
+- Macro de MJ : afficher les propriétés principales des PJs
 
 ## [1.10.0] - 2021-01-16
 ### Ajouté